refactor: move the deployment schema annotations out of the view
The annotation serializers exist only to shape the published spec, so they
sit in `api_v2/openapi_schema.py` rather than in `serializers.py`, where a
request-time import of one would look ordinary. `api_deployment_views.py`
keeps a single decorator.

`deployment_spec_urls.py` now selects mounts from a tuple, so widening the
spec to another endpoint is one entry plus its `@extend_schema`.

The generated spec is unchanged: component names are all that reach it.
