+++ backend/api_v2/deployment_spec_urls.py
@@ -1,27 +1,29 @@
-"""URLconf used only to generate the API deployment OpenAPI spec.
+"""URLconf the published OpenAPI spec is generated against.
 
-``api_v2.execution_urls`` is an included sub-urlconf, so generating against it
-directly yields paths without the prefix it is mounted at — a spec describing
-URLs the server does not serve. The mount is selected out of the served
-urlconf rather than restated here, so a change to where the deployment API is
-mounted moves the generated paths with it.
+Each entry is an included sub-urlconf: generating against one directly yields
+paths without the prefix it is mounted at, i.e. a spec describing URLs the
+server does not serve. The mounts are selected out of the served urlconf
+rather than restated, so moving one moves the generated paths with it.
+
+Widening the spec to another endpoint means annotating its view with
+``@extend_schema`` and adding its urlconf here.
 """
 
 from django.core.exceptions import ImproperlyConfigured
 
 from backend import base_urls
 
-DEPLOYMENT_URLCONF = "api_v2.execution_urls"
+SPEC_URLCONFS = ("api_v2.execution_urls",)
 
 urlpatterns = [
     entry
     for entry in base_urls.urlpatterns
-    if getattr(getattr(entry, "urlconf_name", None), "__name__", None)
-    == DEPLOYMENT_URLCONF
+    if getattr(getattr(entry, "urlconf_name", None), "__name__", None) in SPEC_URLCONFS
 ]
 
-if not urlpatterns:
+missing = set(SPEC_URLCONFS) - {entry.urlconf_name.__name__ for entry in urlpatterns}
+if missing:
     raise ImproperlyConfigured(
-        f"{DEPLOYMENT_URLCONF} is not mounted in backend.base_urls; the API "
-        "deployment spec would be generated for no routes at all."
+        f"{', '.join(sorted(missing))} is not mounted in backend.base_urls; the "
+        "spec would be generated for routes the server does not serve."
     )
